test(db): move discovery tests next to module + hoist seed fixtures
